Add some settings options and fix a problem with the ruler setting.
- The way that the checkbox options was made was not good, but work really well, and that's good enough.

I wanted to write a directive for all type of setting (<select></select> for exemple), but maybe some day,
i need to finish this fast, because i need to learn english for real... yeah,
and during the next week i'm going to study english, i need to learn a lot of words and etc.

I need to sleep now. :D